+++ sdk/core/typespec_client_core/src/stream/mod.rs
@@ -18,8 +18,7 @@ use std::{pin::Pin, task::Context};
 pub const DEFAULT_BUFFER_SIZE: usize = 1024 * 64;
 
 /// Enable a type implementing `AsyncRead` to be consumed as if it were a `Stream` of `Bytes`.
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[async_trait::async_trait]
 pub trait SeekableStream: AsyncRead + Unpin + std::fmt::Debug + Send + Sync + DynClone {
     /// Resets the stream position to the beginning.
     async fn reset(&mut self) -> Result<()>;