+++ sdk/cosmos/azure_data_cosmos/src/clients/cosmos_client.rs
@@ -95,11 +95,6 @@ impl CosmosClient {
                 options.enable_partition_level_circuit_breaker,
             );
 
-        #[allow(
-            clippy::arc_with_non_send_sync,
-            reason = "Wasm32 doesn't include Send, but it's also single-threaded so it's fine"
-        )]
-        // On wasm32 SpawnedTask is !Send; Arc is still correct.
         let pipeline = Arc::new(GatewayPipeline::new(
             endpoint,
             pipeline_core,
@@ -186,11 +181,6 @@ impl CosmosClient {
             options.enable_partition_level_circuit_breaker,
         );
 
-        #[allow(
-            clippy::arc_with_non_send_sync,
-            reason = "Wasm32 doesn't include Send, but it's also single-threaded so it's fine"
-        )]
-        // On wasm32 SpawnedTask is !Send; Arc is still correct.
         let pipeline = Arc::new(GatewayPipeline::new(
             endpoint,
             pipeline_core,