+++ sdk/eventhubs/azure_messaging_eventhubs/src/common/recoverable/claims_based_security.rs
@@ -65,8 +65,7 @@ impl RecoverableClaimsBasedSecurity {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[async_trait::async_trait]
 impl AmqpClaimsBasedSecurityApis for RecoverableClaimsBasedSecurity {
     async fn authorize_path(
         &self,