+++ sdk/servicebus/azure_messaging_servicebus/src/common/authorizer.rs
@@ -416,8 +416,7 @@ mod tests {
         }
     }
 
-    #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-    #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+    #[async_trait::async_trait]
     impl TokenCredential for MockTokenCredential {
         async fn get_token(
             &self,