+++ sdk/storage/azure_storage_blob/tests/streaming.rs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#![cfg(not(target_arch = "wasm32"))]
-
 use azure_core::{
     http::{Body, NoFormat, RequestContent},
     stream::BytesStream,