+++ sdk/storage/azure_storage_blob/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 
 - Renamed `ListBlobsFlatSegmentResponse` to `ListBlobsResponse`.
+- Support for `wasm32-unknown-unknown` has been removed ([#3377](https://github.com/Azure/azure-sdk-for-rust/issues/3377))
 
 ### Bugs Fixed